graphics.library and graphics.hidd LONG to WORD coordinate conversion.
git-svn-id: https://svn.aros.org/svn/aros/trunk@45159 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/graphics/blttemplate.c
@@ -23,13 +23,13 @@ struct bt_render_data
     UBYTE  inverttemplate;
 };
 
-static ULONG blttemplate_render(APTR btr_data, LONG srcx, LONG srcy,
+static ULONG blttemplate_render(APTR btr_data, WORD srcx, WORD srcy,
     	    	    	    	OOP_Object *dstbm_obj, OOP_Object *dst_gc,
     	    	    	    	struct Rectangle *rect, struct GfxBase *GfxBase)
 {
     struct bt_render_data *btrd = btr_data;
-    ULONG   	    	   width  = rect->MaxX - rect->MinX + 1;
-    ULONG		   height = rect->MaxY - rect->MinY + 1;
+    WORD   	    	   width  = rect->MaxX - rect->MinX + 1;
+    WORD		   height = rect->MaxY - rect->MinY + 1;
     UBYTE		   x = srcx + btrd->srcx;
     UBYTE   	    	  *template = btrd->template + btrd->modulo * srcy;
     
@@ -49,13 +49,13 @@ static ULONG blttemplate_render(APTR btr_data, LONG srcx, LONG srcy,
 /*  SYNOPSIS */
 
 	AROS_LHA(PLANEPTR		, source	, A0),
-	AROS_LHA(LONG              	, xSrc		, D0),
-	AROS_LHA(LONG              	, srcMod	, D1),
+	AROS_LHA(WORD              	, xSrc		, D0),
+	AROS_LHA(WORD              	, srcMod	, D1),
 	AROS_LHA(struct RastPort * 	, destRP	, A1),
-	AROS_LHA(LONG              	, xDest		, D2),
-	AROS_LHA(LONG              	, yDest		, D3),
-	AROS_LHA(LONG              	, xSize		, D4),
-	AROS_LHA(LONG              	, ySize		, D5),
+	AROS_LHA(WORD              	, xDest		, D2),
+	AROS_LHA(WORD              	, yDest		, D3),
+	AROS_LHA(WORD              	, xSize		, D4),
+	AROS_LHA(WORD              	, ySize		, D5),
 
 /*  LOCATION */
 	struct GfxBase *, GfxBase, 6, Graphics)
